Add files via upload
This version of the Blofeld Editor expands the editor with GUI elements to edit individual pages of the Blofeld synthesizer.  Included in the main program are subpatches to assist in creating new subpatches and abstractions:  MIDI Monitor (which includes the ability to monitor system exclusive messages) and a Canvas Designer to resize canvases.  Abstractions are included for all Filter 1 pages.  

To do:  Expand to include Filter 2 and filter routing, revise the step sequencer to use arrays.